requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
channels==4.0.0
charset-normalizer==3.1.0
click==8.1.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
distlib==0.3.6
Django==4.2.3
django-environ==0.10.0
djangorestframework==3.14.0
drf-yasg==1.21.6
filelock==3.12.2
frozenlist==1.3.3
idna==3.4
inflection==0.5.1
kombu==5.3.1
multidict==6.0.4
mysqlclient==2.2.0
openai==0.27.8
packaging==23.1
platformdirs==3.8.0
prompt-toolkit==3.0.39
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tqdm==4.65.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
vine==5.0.0
virtualenv==20.23.1
virtualenv-clone==0.5.7
wcwidth==0.2.6
yarl==1.9.2
